Changes for setting modified_enc1 to generated_enc1 when sentence changes and not setting gradients after len1[0] to 0
Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Not setting gradients to 0. Connected with previous commits

Setting modified_enc1 to the generated_enc1 when sentences changes

Setting modified_enc1 to the generated_enc1 when sentences changes

Setting modified_enc1 to the generated_enc1 when sentences changes